// pronti-frontend-main/front/src/app/enums/codigo-auxiliar-rubros-rol.enum.ts
export enum CodigoAuxiliarRubrosRolEnum {
  argumentoDeEgreso = 'AE',
  argumentoDeIngreso = 'AI',
  egresosFijos = 'EF',
  egresosLegales = 'EL',
  otrosEgresos = 'OE',
  egresosPrestamos = 'EP',
  ingresosFijos = 'IF',
  ingresosLegales = 'IL',
  otrosIngresos = 'IO',
  totalesAcumulados = 'TA',
  totalesIndividuales = 'TI',
  provisiones = 'PV'

}